add logic to initialize auctions and set up the max auctions const

+++ contracts/src/models/auction.cairo
@@ -18,14 +18,14 @@ pub struct Auction {
     pub start_price: u256,
     pub floor_price: u256,
     pub is_finished: bool,
-    pub decay_rate: u8
+    pub decay_rate: u64
 }
 
 #[generate_trait]
 impl AuctionImpl of AuctionTrait {
     #[inline(always)]
     fn new(
-        land_location: u64, start_price: u256, floor_price: u256, is_finished: bool, decay_rate: u8
+        land_location: u64, start_price: u256, floor_price: u256, is_finished: bool, decay_rate: u64
     ) -> Auction {
         Auction {
             land_location,
